added some more checks to the bootstrap script

+++ vcl/bootstrap_ansible.sh
@@ -1,28 +1,47 @@
 #!/bin/bash
 
-if ! rpm -qa | grep epel-release; then
-    rpm -ivh http://fedora.mirror.nexicom.net/epel/6/i386/epel-release-6-7.noarch.rpm
+NEW_REPO=no
+
+if [[ $UID -ne 0 ]]; then
+    echo "$0 must be run as root"
+    exit 1
 fi
 
-if ! which ansible; then
-    #yum install -y ansible
-    yum localinstall --nogpg http://packages.serverascode.com/mrepo/custom-centos6-noarch/RPMS.all/ansible-0.5-1.el6.noarch.rpm
+if ! rpm -qa | grep epel-release > /dev/null; then
+    rpm -ivh http://fedora.mirror.nexicom.net/epel/6/i386/epel-release-6-7.noarch.rpm
+    NEW_REPO=yes
 fi
+
+if ! rpm -qa | grep rpmforge > /dev/null; then
+	rpm -ivh http://pkgs.repoforge.org/rpmforge-release/rpmforge-release-0.5.2-2.el6.rf.x86_64.rpm
+	NEW_REPO=yes
+fi
+
+if [ "$NEW_REPO" == "yes" ]; then
+	yum clean all
+	yum makcecache
+fi
+
+if ! rpm -qa ansible > /dev/null ; then
+    yum install -y ansible
+fi
+
+if [ ! -e ./ansible_hosts ]; then
 
 cat << EOANSIBLE_HOSTS > ./ansible_hosts
 [vcl]
 127.0.0.1
 EOANSIBLE_HOSTS
 
+fi
+
 export ANSIBLE_HOSTS=`pwd`/ansible_hosts
 
-mkdir /root/.ssh
-
-chmod 700 /root/.ssh
-/usr/bin/ssh-keygen -q -t dsa -C '' -N '' -f /home/vagrant/.ssh/id_dsa
-chown vagrant:vagrant /home/vagrant/.ssh/id_dsa*
-cat /home/vagrant/.ssh/id_dsa.pub >> /root/.ssh/authorized_keys
-
-chmod 600 /root/.ssh/authorized_keys
-
-ansible -m ping 127.0.0.1
+if [ ! -e /root/.ssh/authorized_keys ]; then
+	mkdir /root/.ssh
+	chmod 700 /root/.ssh
+	/usr/bin/ssh-keygen -q -t dsa -C '' -N '' -f /home/vagrant/.ssh/id_dsa
+	chown vagrant:vagrant /home/vagrant/.ssh/id_dsa*
+	cat /home/vagrant/.ssh/id_dsa.pub >> /root/.ssh/authorized_keys
+	chmod 600 /root/.ssh/authorized_keys
+fi
